Update ReadMe
I use this quite a bit and used to trip up on the setup instructions so I've added the exact steps I do when I spin this up for a new website. Once the steps I've added have been done the 'vagrant up' command works perfectly. 

I find that without the folders added the command 'vagrant up' doesn't work as it's looking the puppet content and it fails to create/find a  'www' folder too.